Add etch-a-sketch functionality
This commit adds the etch-a-sketch (yeah, all of it. I definitely need
to get better at atomic commits). It uses for loops to create a series
of flexbox gridDivs nested inside flexbox rowDivs. The gridDivs are
also given an event listener to turn them black when the mouse passes
over them. The initial grid is 16 by 16, but a button at the top clears
the container and prompts the user for a new grid size, which gets
passed into the createDivs function as its parameters.