Modify mouseenter event to increment alpha values
This commit modifies the mouseenter event in the createDivs function to
increment the alpha value of a square by 0.1 every time the mouse
passes over it. This one was a little more difficult, as it required
a pretty complicated regex to read the current alpha value of a square,
and to be honest, I still don't feel like I understand how regex works
very well. But it does work!