Add button to toggle border visibility
This commit adds a button that will toggle the border for the grid
container and all its child elements on and off. It also adds a couple
small tweaks - it slightly alters the createDivs function to use one
parameter rather than 2 identical parameters, and it fixes a problem
with the mouseenter event listener where the color of each gridDiv
would stop being randomized once it's alpha value reached 1.